+++ tests/Agent/IntegrationTests/IntegrationTests/AgentFeatures/ThreadProfileStressTests.cs
@@ -63,15 +63,20 @@ namespace NewRelic.Agent.IntegrationTests.AgentFeatures
             fixture.Initialize();
         }
 
-        [SkipUntilDateFact("2019-09-01", "This test is flaky and fails on the server almost all the time.")]
+        [Fact]
         public void Test()
         {
             NrAssert.Multiple(
                 () => Assert.Contains(@"""OTHER"":[[[""Native"",""Function Call"",0]", _threadProfileString),
                 () => Assert.Contains(@"[""ThreadProfileStressTest.Program"",""Main"",0]", _threadProfileString),
-                () => Assert.Contains(@"[""ThreadProfileStressTest.Program"",""DoTheThing"",0]", _threadProfileString),
-                () => Assert.Contains(@"[""ThreadProfileStressTest.Program"",""DoWork"",0]", _threadProfileString)
+                () => Assert.Contains(@"[""ThreadProfileStressTest.Program"",""DoTheThing"",0]", _threadProfileString)
             );
+
+            // This assertion has been commented out since the purpose of this test is to verify that the profiler doesn't CRASH
+            // when threads are frequently created/destroyed. In this stress scenario we are unable to capture stack traces for finalized
+            // threads. Ideally we would get atleast one sample with the 'DoWork' method, but the purpose of the test is to verify we don't crash
+
+            // Assert.Contains(@"[""ThreadProfileStressTest.Program"",""DoWork"",0]", _threadProfileString);
         }
     }
 }