+++ tests/Agent/IntegrationTests/IntegrationTests/RestSharp/RestSharpClientInstrumentationCAT.cs
@@ -27,6 +27,7 @@ namespace NewRelic.Agent.IntegrationTests.RestSharp
                     var configPath = fixture.DestinationNewRelicConfigFilePath;
                     var configModifier = new NewRelicConfigModifier(configPath);
 
+                    configModifier.SetLogLevel("finest");
                     configModifier.EnableCat();
                     configModifier.ForceTransactionTraces();
                 },